Payment Success + Failure components moved to submitted-card folder
